requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
opencv-python>=4.8.0
easyocr>=1.7.0
numpy>=1.24.0
Pillow>=10.0.0
pyodbc>=5.0.0
httpx>=0.25.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
